requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
cached-property==1.5.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
cssselect==1.0.3
demjson==2.2.4
diff-match-patch==20121119
Django==2.1.1
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-import-export==1.0.1
django-mongonaut==0.2.21
django-oauth-toolkit==1.2.0
djangorestframework==3.8.2
djangorestframework-oauth==1.1.0
docopt==0.6.2
et-xmlfile==1.0.1
facebook-wda==0.2.1
future==0.16.0
gitdb2==2.0.4
GitPython==2.1.11
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
interfaces==0.0.4
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jsonschema==2.6.0
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
mongoengine==0.9.0
nose==1.3.7
numpy==1.15.1
oauthlib==2.1.0
odfpy==1.3.6
opencv-contrib-python==3.2.0.7
openpyxl==2.5.8
parsel==1.5.0
Pillow==5.2.0
poco==0.96.1
pprint==0.1
pyaml==16.12.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyGithub==1.39
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pymongo==3.7.1
PyMySQL==0.9.2
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-gitlab==1.6.0
python-xlib==0.23
pytz==2018.5
pywinauto==0.6.3
PyYAML==3.13
queuelib==1.5.0
requests==2.18.4
Scrapy==1.5.1
service-identity==17.0.0
six==1.11.0
smmap2==2.0.4
SQLAlchemy==1.2.12
svn==0.3.44
tablib==0.12.1
texttable==0.9.1
Twisted==18.7.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
w3lib==1.19.0
websocket-client==0.53.0
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.5.0
